Relocate paragaph
