--- src/styles/paragraph.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const Paragraph = styled.p`
-  margin: 0;
-`;
