+++ src/styles/text-styles.ts
@@ -9,3 +9,7 @@ export const H1 = styled.h1`
 export const H3 = styled.h3`
   margin: 0;
 `;
+
+export const Paragraph = styled.p`
+  margin: 0;
+`;
